158_disambiguator/requirements.txt
werkzeug
uwsgi
clint
gensim
nltk
pandas==0.24.2
Flask
flasgger
psycopg2-binary
